# src/server/openapi.yaml
openapi: 3.0.3
info:
  title: JA Accountable API
  description: API for searching news articles and browsing extracted entities with classification data.
  version: 1.0.0

servers:
  - url: /api/v1
    description: API v1

paths:
  /articles:
    get:
      summary: Search articles
      description: |
        Search for articles using full-text search (PostgreSQL FTS on title + full_text),
        with optional filtering by date range and sorting options.
      operationId: searchArticles
      tags:
        - Articles
      parameters:
        - name: q
          in: query
          description: Full-text search query (uses PostgreSQL FTS on title + full_text)
          required: false
          schema:
            type: string
          example: corruption

        - name: from_date
          in: query
          description: Articles published on or after this date
          required: false
          schema:
            type: string
            format: date-time
          example: '2024-01-01T00:00:00Z'

        - name: to_date
          in: query
          description: Articles published on or before this date
          required: false
          schema:
            type: string
            format: date-time
          example: '2024-12-31T23:59:59Z'

        - name: include_full_text
          in: query
          description: Include full article text in response
          required: false
          schema:
            type: boolean
            default: false

        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1

        - name: page_size
          in: query
          description: Results per page (max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: sort
          in: query
          description: Sort field
          required: false
          schema:
            type: string
            enum:
              - relevance
              - published_date
            default: relevance

        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc

      responses:
        '200':
          description: Successful search response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleSearchResponse'

        '400':
          description: Invalid query parameters (bad date format, page < 1, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '500':
          description: Database/server errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /articles/{public_id}/related:
    get:
      summary: Get related articles
      description: |
        Returns articles related to the given article, ranked by number of shared
        entities, wrapped in a `{ "articles": [...] }` envelope. Articles from any
        news source may appear. Returns an empty `articles` array when the article
        has no entity links or no other articles share its entities.
      operationId: getRelatedArticles
      tags:
        - Articles
      parameters:
        - name: public_id
          in: path
          required: true
          description: Public UUID of the source article
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000

        - name: limit
          in: query
          required: false
          description: Maximum number of related articles to return (default 5, max 20)
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5

      responses:
        '200':
          description: Related articles (empty array if none found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelatedArticlesResponse'

        '422':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '500':
          description: Database/server errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /articles/{public_id}:
    get:
      summary: Get article by ID
      description: Retrieve a single article by its public UUID, with full text, entities, and classifications.
      operationId: getArticle
      tags:
        - Articles
      parameters:
        - name: public_id
          in: path
          required: true
          description: Public UUID of the article
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000

      responses:
        '200':
          description: Article found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'

        '404':
          description: Article not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '422':
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '500':
          description: Database/server errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /entities:
    get:
      summary: List entities
      description: |
        List entities (people, organizations, government bodies) extracted from
        articles, with article counts and recency data.

        Supports sorting by most recent appearance or by frequency, with optional
        time-window scoping via the `since` parameter. When `since` is provided,
        both article_count and last_seen_date reflect only articles within that window.
      operationId: listEntities
      tags:
        - Entities
      parameters:
        - name: sort
          in: query
          description: |
            Sort mode:
            - `latest`: ordered by most recent article appearance
            - `most_found`: ordered by article count (most appearances first)
          required: false
          schema:
            type: string
            enum:
              - latest
              - most_found
            default: latest

        - name: since
          in: query
          description: |
            Scope results to articles published on or after this date.
            When provided, both article_count and last_seen_date reflect
            only articles within this time window.
          required: false
          schema:
            type: string
            format: date
          example: '2024-01-01'

        - name: page
          in: query
          description: Page number (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1

        - name: page_size
          in: query
          description: Results per page (max 100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityListResponse'

        '400':
          description: Invalid query parameters (bad date format, page < 1, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

        '500':
          description: Database/server errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    # ── Article Search ──────────────────────────────────────────────

    ArticleSearchResponse:
      type: object
      required:
        - items
        - total
        - page
        - pageSize
        - pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        total:
          type: integer
          description: Total number of matching results
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of results per page
          example: 20
        pages:
          type: integer
          description: Total number of pages
          example: 8

    Article:
      type: object
      required:
        - id
        - url
        - title
        - section
        - newsSource
        - publishedDate
        - snippet
        - entities
        - classifications
      properties:
        id:
          type: string
          format: uuid
          description: Public article identifier (UUID)
          example: 550e8400-e29b-41d4-a716-446655440000
        url:
          type: string
          format: uri
          description: Original article URL
          example: https://jamaica-gleaner.com/article/...
        title:
          type: string
          description: Article title
          example: Government Minister Under Investigation
        section:
          type: string
          description: Article section/category
          example: news
        newsSource:
          type: string
          enum:
            - JAMAICA_GLEANER
            - JAMAICA_OBSERVER
          description: News outlet identifier
          example: JAMAICA_GLEANER
        publishedDate:
          type: string
          format: date-time
          nullable: true
          description: Article publication date
          example: '2024-01-15T10:30:00Z'
        snippet:
          type: string
          nullable: true
          description: Highlighted excerpt with search terms marked
          example: '...highlighted excerpt with <mark>search terms</mark>...'
        entities:
          type: array
          items:
            type: string
          description: List of extracted entity names
          example:
            - John Smith
            - Ministry of Finance
        classifications:
          type: array
          items:
            $ref: '#/components/schemas/Classification'
          description: List of article classifications with confidence scores
        fullText:
          type: string
          nullable: true
          description: Full article text (only included if include_full_text=true)

    Classification:
      type: object
      required:
        - classifierType
        - confidenceScore
      properties:
        classifierType:
          type: string
          description: Type of classification
          example: CORRUPTION
        confidenceScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for the classification
          example: 0.92

    RelatedArticlesResponse:
      type: object
      required:
        - articles
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'

    # ── Entities ────────────────────────────────────────────────────

    EntityListResponse:
      type: object
      required:
        - items
        - total
        - page
        - pageSize
        - pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EntitySummary'
        total:
          type: integer
          description: Total number of matching results
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of results per page
          example: 20
        pages:
          type: integer
          description: Total number of pages
          example: 8

    EntitySummary:
      type: object
      required:
        - name
        - normalizedName
        - articleCount
        - lastSeenDate
      properties:
        name:
          type: string
          description: Display name of the entity as originally extracted
          example: Ministry of Finance
        normalizedName:
          type: string
          description: Normalized canonical name used for deduplication
          example: ministry_of_finance
        articleCount:
          type: integer
          description: Number of articles this entity appears in (scoped by time window if since is provided)
          example: 32
        lastSeenDate:
          type: string
          format: date-time
          description: Most recent article publication date featuring this entity (scoped by time window if since is provided)
          example: '2024-02-14T09:15:00Z'

    # ── Shared ──────────────────────────────────────────────────────

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: BAD_REQUEST
        message:
          type: string
          description: Human-readable error message
          example: Invalid date format for from_date parameter